requirements.txt
anthropic==0.86.0
PyNaCl==1.5.0
python-dotenv==1.2.1
beautifulsoup4==4.14.3
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
curl_cffi==0.13.0  # optional — Jobstreet Chrome impersonation
httpx==0.28.1
certifi==2026.2.25

# dev
pytest==8.4.2